Flytter dokumentasjonshåntering ut fra Dokumentasjon.tsx
