+++ src/frontend/context/AppContext.ts
@@ -14,9 +14,11 @@ import {
 } from '@navikt/familie-typer';
 
 import Miljø, { basePath } from '../../shared-utils/Miljø';
+import { hentVedleggOppsummering } from '../components/Felleskomponenter/VedleggOppsummering/vedleggOppsummeringUtils';
 import { DinLivssituasjonSpørsmålId } from '../components/SøknadsSteg/DinLivssituasjon/spørsmål';
 import { LocaleType } from '../typer/common';
 import { IKontoinformasjon } from '../typer/kontoinformasjon';
+import { Dokumentasjonsbehov } from '../typer/kontrakt/dokumentasjon';
 import { ESivilstand, ESøknadstype, TilRestLocaleRecord } from '../typer/kontrakt/generelle';
 import { IKvittering } from '../typer/kvittering';
 import { IMellomlagretBarnetrygd } from '../typer/mellomlager';
@@ -27,6 +29,7 @@ import { ITekstinnhold } from '../typer/sanity/tekstInnhold';
 import { initialStateSøknad, ISøknad } from '../typer/søknad';
 import { InnloggetStatus } from '../utils/autentisering';
 import { mapBarnResponsTilBarn } from '../utils/barn';
+import { hentRelevateDokumentasjoner } from '../utils/dokumentasjon';
 import { plainTekstHof } from '../utils/sanity';
 
 import { preferredAxios } from './axios';
@@ -361,6 +364,20 @@ const [AppProvider, useApp] = createUseContext(() => {
         };
     };
 
+    const relevateDokumentasjoner = hentRelevateDokumentasjoner(søknad.dokumentasjon);
+
+    const relevateDokumentasjonerUtenAnnenDokumentasjon = relevateDokumentasjoner.filter(
+        dokumentasjon =>
+            dokumentasjon.dokumentasjonsbehov !== Dokumentasjonsbehov.ANNEN_DOKUMENTASJON
+    );
+
+    const brukerHarVedleggskrav = relevateDokumentasjonerUtenAnnenDokumentasjon.length > 0;
+
+    const vedleggOppsummering = hentVedleggOppsummering(
+        relevateDokumentasjonerUtenAnnenDokumentasjon,
+        søknad
+    );
+
     return {
         axiosRequest,
         sluttbruker,
@@ -393,6 +410,10 @@ const [AppProvider, useApp] = createUseContext(() => {
         tilRestLocaleRecord,
         flettefeltTilTekst,
         kontoinformasjon,
+        relevateDokumentasjoner,
+        relevateDokumentasjonerUtenAnnenDokumentasjon,
+        brukerHarVedleggskrav,
+        vedleggOppsummering,
     };
 });
